backend/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.9
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.6.1
httpx==0.26.0
SpeechRecognition==3.10.0
pydub==0.25.1
